Increase precision of m_last
double limited us to 2e-8 s precision per sample in the window, long double
gets us to a window independent 3e-7 s (which only can be observed
by looking at the internal variable).

+++ src/TimestampEstimator.cpp
@@ -158,10 +158,10 @@ base::Time TimestampEstimator::update(base::Time time)
         m_last = m_last + period;
         if (current - m_last < 0)
             throw std::logic_error("base time is after current sample");
-        m_min_offset = std::min(m_min_offset, current - m_last);
+        m_min_offset = std::min(m_min_offset, current - (double)m_last);
     }
 
-    return base::Time::fromSeconds(m_last);
+    return base::Time::fromSeconds((double)m_last);
 }
 
 void TimestampEstimator::updateLoss()
